# whatsappclone/src/main/java/com/zel/whatsappclone/message/MessageController.java
package com.zel.whatsappclone.message;

import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;

import static org.springframework.http.HttpStatus.ACCEPTED;
import static org.springframework.http.HttpStatus.CREATED;

@RestController
@RequestMapping("/api/v1/messages")
@RequiredArgsConstructor
@Tag(name = "Message")
public class MessageController {
    private final MessageService service;

    @PostMapping
    @ResponseStatus(CREATED)
    public void saveMessage(@RequestBody MessageRequest request){
        service.saveMessage(request);
    }

    @PostMapping(value = "/upload-media", consumes = "multipart/form-data")
    @ResponseStatus(CREATED)
    public void uploadMedia(@RequestParam("chat-id") String chatId,
                            @Parameter
                            @RequestParam("file")MultipartFile file,
                            Authentication authentication){
        service.uploadMediaMessage(chatId, file, authentication);
    }

    @PatchMapping
    @ResponseStatus(ACCEPTED)
    public void setMessagesToSeen(@RequestParam("chat-id") String chatId, Authentication authentication){
        service.setMessagesToSeen(chatId ,authentication);
    }

    @GetMapping("/chat/{chat-id}")
    public ResponseEntity<List<MessageResponse>> findChatMessages(@PathVariable("chat-id") String chatId){
        return ResponseEntity.ok(service.findChatMessages(chatId));
    }
}